requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
python-docx>=1.1.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
